# docs/swagger.yaml
definitions:
  http_interfaces_authentication.AuthenticationResponse:
    properties:
      token:
        type: string
    type: object
  http_interfaces_authentication.CreateAuthenticationRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  http_interfaces_customer.CreateCustomerRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
  http_interfaces_customer.CustomerResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  http_interfaces_customer.UpdateCustomerRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  http_interfaces_favorite.AddFavoriteRequest:
    properties:
      product_id:
        type: integer
    required:
    - product_id
    type: object
  http_interfaces_favorite.FavoriteResponse:
    properties:
      customer_id:
        type: integer
      id:
        type: integer
      product:
        type: string
      product_id:
        type: integer
    type: object
info:
  contact: {}
  description: API para favoritar produtos
  title: Favorite API
  version: "1.0"
paths:
  /authenticate:
    post:
      consumes:
      - application/json
      description: Creates a JWT token for the user
      parameters:
      - description: Authentication data
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_authentication.CreateAuthenticationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_authentication.AuthenticationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create authentication token
      tags:
      - Authentication
  /customer:
    post:
      consumes:
      - application/json
      description: Creates a favorite for a given customer and product
      parameters:
      - description: Customer data
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_customer.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add customer product
      tags:
      - Customer
  /customer/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a customer by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete customer
      tags:
      - Customer
    get:
      consumes:
      - application/json
      description: Retrieves a customer's information by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get customer by ID
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Updates a customer's information by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_customer.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update customer
      tags:
      - Customer
  /customer/{id}/favorites:
    get:
      consumes:
      - application/json
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http_interfaces_favorite.FavoriteResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar favoritos do cliente
      tags:
      - Favorites
    post:
      consumes:
      - application/json
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Produto favorito
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_favorite.AddFavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http_interfaces_favorite.FavoriteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Adicionar produto aos favoritos do cliente
      tags:
      - Favorites
  /customer/{id}/favorites/{productId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product ID
        in: path
        name: productId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remover produto dos favoritos do cliente
      tags:
      - Favorites
  /customers:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of customers
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Customer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"